requirements.txt
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
streamlit>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
